move widths to separate file
